requirements.txt
streamlit>=1.32.0
web3>=6.15.0
eth-brownie>=1.20.0
pandas>=2.0.0